Creating page

+++ src/main/java/iducs/springboot/board/service/QuestionService.java
@@ -7,7 +7,7 @@ import iducs.springboot.board.domain.Question;
 public interface QuestionService {
 	Question getQuestionById(long id); // primary key인 id 값을 가진 질문 조회
 	List<Question> getQuestions();
-	List<Question> getQuestions(Long pageNo); // 모든 질문  조회
+	List<Question> getQuestions(int pageNo); // 모든 질문  조회
 	
 	List<Question> getQuestionsByUser(String name); // name으로 조회
 	List<Question> getQuestionsByPage(int index, int size); // 페이지로 조회
